requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
tomli==2.2.1
typing_extensions==4.12.2
Werkzeug==3.1.3
